cleanup more warnings

+++ src/lexer/nmodl_utils.cpp
@@ -97,7 +97,7 @@ SymbolType prime_symbol(std::string text, PositionType& pos) {
 
     auto prime_name = new ast::String(text);
     assert(order <= std::numeric_limits<int>::max());
-    auto prime_order = new ast::Integer(order, nullptr);
+    auto prime_order = new ast::Integer(static_cast<int>(order), nullptr);
     ast::PrimeName value(prime_name, prime_order);
     value.set_token(token);
     return Parser::make_PRIME(value, pos);
